requirements.txt
numpy>=1.20.0
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.4.0
scipy>=1.7.0
moviepy>=1.0.3
cryptography